Finished chest model, fixed Player mouse buttons and using items/blocks, fixed torch look and placing bugs, fixed interacting with hotbar with game is paused, made items drop away from the player, moved mineshaft pieces up ont he y axis, finished shaft piece, finished crossing piece, added ladder, added Chunk.isInchunk with ignore y varient, fixed multi mesh blocks in hotbar

+++ Assets/VoxelEngine/Items/ItemSword.cs
@@ -2,9 +2,9 @@
 
     public class ItemSword : Item {
 
-        public float damageAmount;
+        public int damageAmount;
 
-        public ItemSword(int id, float damage) : base(id) {
+        public ItemSword(int id, int damage) : base(id) {
             this.damageAmount = damage;
         }
     }
